requirements.txt
attrs
Django
sqlparse
